MATH-3: renamed main package
