+++ src/main/java/io/github/orionlibs/math/NewClass.java
@@ -1,4 +1,4 @@
-package io.github.orionlibs.orionmath;
+package io.github.orionlibs.math;
 
 import lombok.NoArgsConstructor;
